+++ src/lib.rs
@@ -239,7 +239,7 @@ mod tests {
 
     #[test]
     fn monte_carlo() {
-        let mut tree = Tree::new(BoardSize::_11);
+        let mut tree = Tree::new(Game::new_game(BoardSize::_11, None));
         let duration = Duration::from_secs(1);
         let depth = 80;
         let (loops, _plays) = tree.monte_carlo_tree_search(duration, depth);
@@ -249,7 +249,7 @@ mod tests {
     #[ignore = "takes too long"]
     #[test]
     fn monte_carlo_long() {
-        let mut tree = Tree::new(BoardSize::_11);
+        let mut tree = Tree::new(Game::new_game(BoardSize::_11, None));
         let duration = Duration::from_secs(10);
         let depth = 40;
         let (loops, _plays) = tree.monte_carlo_tree_search(duration, depth);